requirements.txt
gym==0.25.2
matplotlib
numpy
stable-baselines3
neat-python
shimmy>=2.0